Changes to Jump, Gap Hover, added Random Forces
Jumps are now stored via an enum instead of boolean variables for consistency and ease of use

Gap Hover has been fixed so that the relative Y position changes to accurately reflect where the player would be in relation to the last detected platform

Random forces are sometimes applied to the player to keep the player alternating up and down. The amount of force used should be small enough not to make a difference in gameplay.